Seven ways to tell the search where to look, and one to clean up after it
Version 0.10.2. The batch is mostly about giving the user a way to state
something about a book that the tool would otherwise have to guess at, plus a
housekeeping mode for the traces it leaves behind.

--cleanup undoes what earlier runs left in a folder: leftover temporary files,
.debug.log files, interrupted batches' progress files, .missing-marks name
tags, and .bak backups - the last only where the file they back up is beside
them and probes to the same length, so it can never throw away the only copy of
anything. An orphaned .abchapterize.orig is the audiobook itself and gets put
back rather than deleted. Everything is decided before anything is done, which
is what lets the prompt describe the whole plan and what makes the ordering
expressible: untagging has to be planned before the backups, since a tagged book
is what will be sitting next to Book.m4b.bak by the time they are looked at.

--chapter-count closes the one hole nothing else structurally can: a chapter
missing after the last one found has nothing above it to compare against.
--trailing-scan brute-forces that by transcribing the whole tail on spec; told
the count, the run knows which numbers are owed and stops when they turn up.

The silence threshold is now measured per file instead of fixed at -35 dBFS,
with --noise-floor to override it. Calibrated against the fourteen-book corpus
on 2026-08-05 (eight 20 s excerpts each, 50 ms RMS frames): the narrowest
margins -35 actually had were 9.9 dB under the quietest sustained speech (I
Shall Wear Midnight, p75 -25.1) and 15.1 dB over the loudest room tone (The
Philosopher's Stone, p5 -50.1), which is where the two headroom constants come
from, each rounded down for slack. Every one of the fourteen still resolves to
exactly -35, verified live on eight of them.

--chapter-phrase none is for books that announce a chapter by speaking its
number alone. What makes it safe is taken from Whisper's own segmentation: a
segment that is a number start to finish is a number spoken between two pauses,
which an ordinary number in prose never is. It leans entirely on the number
sequence to reject a false one, hence the refusal to combine with
--ignore-chapter-numbers.

--verify --fix nudges a confirmed-but-misplaced mark onto its announcement
instead of leaving a full --force re-run as the only way to move it. Bounded at
both ends: under a quarter of a second is not worth remuxing a book for, and
over 30 seconds is not a mark that drifted.

--min-silence-length 0 drops silence-triggered probing entirely, leaving the
jingles. Every VAD region then becomes a candidate, since the dedup that
normally drops one led by a silence has no silence candidate left to dedup
against. The scan itself still runs - marks are placed and refined against it -
and a clip that yields chapter 1 at 0:03:57.46 by default yields it at exactly
0:03:57.46 here.

Also extracts the .missing-marks naming rules out of FileProcessor into
MissingMarksTag, now that three separate paths write and read that tag.

+++ src/Audio/IAudioSource.cs
@@ -32,7 +32,7 @@ public interface IAudioSource
     /// <param name="ct">Cancellation token.</param>
     /// <returns>All detected silence periods in chronological order.</returns>
     Task<List<Silence>> DetectSilencesAsync(
-        string file, double durationSeconds, double minSilenceSeconds, int noiseDb,
+        string file, double durationSeconds, double minSilenceSeconds, double noiseDb,
         Action<double>? progress, string? inputDecoder, CancellationToken ct);
 
     /// <summary>
@@ -70,7 +70,7 @@ public interface IAudioSource
     /// <param name="ct">Cancellation token.</param>
     /// <returns>All detected silence periods in chronological order.</returns>
     Task<List<Silence>> DetectSilencesAndStreamPcmAsync(
-        string file, double durationSeconds, double minSilenceSeconds, int noiseDb,
+        string file, double durationSeconds, double minSilenceSeconds, double noiseDb,
         Func<IAsyncEnumerable<float[]>, CancellationToken, Task> consumePcm,
         Action<double>? progress, string? inputDecoder, CancellationToken ct);
 }
